速度向上の試行錯誤

+++ VRP/cpp/src/construct.hpp
@@ -5,5 +5,5 @@
 using std::vector;
 using std::pair;
 
-vector<vector<int>> InsertConstruct(vector<vector<double>> dis_mat,vector<int> weights,
+vector<vector<int>> InsertConstruct(vector<vector<float>> dis_mat,vector<int> weights,
                                     int truck_capacity,int truck_size,vector<pair<int,int>>& truck_ids);
